+++ src/main/java/de/thowl/tnt/core/AuthenticationServiceImpl.java
@@ -72,10 +72,10 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 	@Scheduled(fixedRate = 60000)
 	public void cleanupExpiredSessions() {
 
-		log.debug("entering cleanupExpiredSessions");
-
 		Date now;
 		List<Session> expired;
+
+		log.debug("entering cleanupExpiredSessions");
 
 		now = new Date();
 		expired = sessions.findByExpiresAtBefore(now);
@@ -92,10 +92,10 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 	@Override
 	public boolean validateEmail(String email) {
 
-		log.debug("entering validateEmail");
-
 		String regex;
 		boolean result;
+
+		log.debug("entering validateEmail");
 
 		if (null == email)
 			return false;
@@ -115,10 +115,10 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 	@Override
 	public boolean validatePassword(String password) {
 
-		log.debug("entering validatePassword");
-
 		String regex;
 		boolean result;
+
+		log.debug("entering validatePassword");
 
 		if (null == password)
 			return false;
@@ -157,11 +157,11 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 	@Override
 	public boolean validateSession(AccessToken token, String username) {
 
-		log.debug("entering validateSession");
-
 		Session session;
 		User user = users.findByUsername(username);
 		boolean result;
+
+		log.debug("entering validateSession");
 
 		if (token == null) {
 			log.error("token was null");
@@ -173,8 +173,7 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 			log.error("a session could not be found");
 			return false;
 		}
-		// String test = session.getAuthToken();
-		// users.findByApiToken(test);
+
 		user = users.findByUsername(username);
 		if (user == null) {
 			log.error("user: {} could not be found", username);
@@ -197,9 +196,9 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 	public void register(String firstname, String lastname, String username,
 			String email, String password) throws DuplicateUserException {
 
-		log.debug("entering register");
-
 		User usr;
+
+		log.debug("entering register");
 
 		if (this.users.findByEmail(email) != null)
 			throw new DuplicateUserException("A User with this Email already exists");
@@ -228,21 +227,19 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 	 */
 	private boolean checkPassword(User user, String password) {
 
-		log.debug("entering checkPassword");
-
 		String bHash;
 		boolean result;
+
+		log.debug("entering checkPassword");
 
 		if (null == password || password.isBlank())
 			return false;
 
 		bHash = user.getPassword();
-
 		log.debug("Comparing Form-password with BCrypt-hash");
 		result = encoder.matches(password, bHash);
 
 		log.debug("checkPassword(user: {}, password: {}) returned: {}", user.toString(), password, result);
-
 		return result;
 	}
 
@@ -254,12 +251,12 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 	 */
 	private AccessToken createSession(User user) {
 
-		log.debug("entering createSession");
-
 		AccessToken token;
 		UUID uuid;
 		Calendar calendar;
 		Date expiryTime;
+
+		log.debug("entering createSession");
 
 		calendar = Calendar.getInstance();
 		calendar.add(Calendar.MINUTE, 30);
@@ -285,9 +282,9 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 	@Override
 	public AccessToken login(String email, String password) throws InvalidCredentialsException {
 
-		log.debug("entering login");
-
 		User user;
+
+		log.debug("entering login");
 
 		if (email == null || password == null) {
 			log.error("One or more params were left empty");
@@ -318,6 +315,8 @@ public class AuthenticationServiceImpl implements AuthenticationService {
 			String email, String password) throws NullUserException {
 
 		User user;
+
+		log.debug("entering updateUser");
 
 		user = users.findById(id).orElseThrow(() -> new NullUserException("User not found"));
