--- src/main/java/de/thowl/tnt/core/services/NoteSearchService.java
@@ -1,35 +1,0 @@
-/*
- *  TnT, Todo's 'n' Texts
- *  Copyright (C) 2023  <name of author>
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package de.thowl.tnt.core.services;
-
-import java.util.List;
-
-import de.thowl.tnt.storage.entities.Note;
-import de.thowl.tnt.storage.entities.NoteKategory;
-
-public interface NoteSearchService {
-    /**
-     * Perform search logic here based on query and filter
-     * 
-     * 
-     * @param query search term
-     * @param filter parameter of type "NoteKategory"
-     */
-    public List<Note> searchNotes(String query, NoteKategory filter);
-    
-}
