+++ src/main/java/de/thowl/tnt/web/NotesController.java
@@ -78,9 +78,9 @@ public class NotesController {
 			@SessionAttribute(name = "token", required = false) AccessToken token,
 			@PathVariable("username") String username, NoteForm form, Model model) {
 
-		log.info("entering doAddNote (POST-Method: /u/{}/notes)", username);
+		String referer;
 
-		String referer = request.getHeader("Referer");
+		log.info("entering doAddNote (POST-Method: /u/{}/notes)", username);
 
 		// Prevent unauthrised access / extend session
 		if (!this.authsvc.validateSession(token, username))
@@ -99,6 +99,7 @@ public class NotesController {
 		this.notessvc.addNote(username, form.getTitle(), form.getSubtitle(), form.getContent(),
 				fileContent, mimeType, form.getKategory(), form.getTags());
 
+		referer = request.getHeader("Referer");
 		return "redirect:" + referer;
 	}
 
@@ -111,13 +112,16 @@ public class NotesController {
 	public String showEditPage(@SessionAttribute(name = "token", required = false) AccessToken token,
 			@PathVariable("username") String username, NoteForm form, Model model,
 			HttpSession httpSession) {
+
+		Note note;
+
 		log.info("entering showEditPage (POST-Method: /u/{}/notes/edit)", username);
 
 		// Prevent unauthrised access / extend session
 		if (!this.authsvc.validateSession(token, username))
 			throw new ForbiddenException("Unathorised access");
 
-		Note note = this.notessvc.getNote(form.getId());
+		note = this.notessvc.getNote(form.getId());
 
 		model.addAttribute("editing", true);
 		model.addAttribute("noteTitle", note.getName());
@@ -136,13 +140,18 @@ public class NotesController {
 	public String doEditNote(@SessionAttribute(name = "token", required = false) AccessToken token,
 			@PathVariable("username") String username, NoteForm form, Model model,
 			HttpSession httpSession) {
+
+		byte[] fileContent;
+		String mimeType;
+
 		log.info("entering doAddNote (POST-Method: /u/{}/notes)", username);
 
 		if (!this.authsvc.validateSession(token, username))
 			throw new ForbiddenException("Unathorised access");
 
-		byte[] fileContent = null;
-		String mimeType = "text/markdown";
+		// Fallback values in case no file was uploaded.
+		fileContent = null;
+		mimeType = "application/octet-stream";
 
 		try {
 			fileContent = form.getFile().getBytes();
@@ -167,9 +176,9 @@ public class NotesController {
 			@SessionAttribute(name = "token", required = false) AccessToken token,
 			@PathVariable("username") String username, NoteForm form, Model model) {
 
-		log.info("entering doDeleteNote (DELETE-Method: /u/{}/notes)", username);
+		String referer;
 
-		String referer = request.getHeader("Referer");
+		log.info("entering doDeleteNote (DELETE-Method: /u/{}/notes)", username);
 
 		// Prevent unauthrised access / extend session
 		if (!this.authsvc.validateSession(token, username))
@@ -177,6 +186,7 @@ public class NotesController {
 
 		this.notessvc.delete(form.getId());
 
+		referer = request.getHeader("Referer");
 		return "redirect:" + referer;
 	}
 }
