+++ src/main/java/de/thowl/tnt/web/ShareController.java
@@ -26,7 +26,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import de.thowl.tnt.core.services.NotesService;
-import de.thowl.tnt.core.services.ShareService;
 import de.thowl.tnt.storage.entities.Note;
 import de.thowl.tnt.web.forms.NoteForm;
 import jakarta.servlet.http.HttpServletRequest;
@@ -37,20 +36,17 @@ import lombok.extern.slf4j.Slf4j;
 public class ShareController {
 
 	@Autowired
-	private ShareService sharesvc;
-
-	@Autowired
 	private NotesService notesvc;
 
 	@RequestMapping(value = "/share/{uuid}", method = RequestMethod.GET)
 	public String showSharePage(@PathVariable("uuid") String uuid, Model model) {
 
-		log.info("entering showSharePage (GET-Method: /share)");
-
 		long id;
 		Note note;
 
-		id = this.sharesvc.getSharedNote(uuid).getNote().getId();
+		log.info("entering showSharePage (GET-Method: /share)");
+
+		id = this.notesvc.getSharedNote(uuid).getNote().getId();
 		note = this.notesvc.getNote(id);
 
 		model.addAttribute("note", note);
@@ -59,17 +55,16 @@ public class ShareController {
 	}
 
 	@RequestMapping(value = "/share", method = RequestMethod.POST)
-	public String addSharedNote(HttpServletRequest request ,NoteForm form, Model model) {
-
-		log.info("entering addSharedNote (Post-Method: /share)");
+	public String addSharedNote(HttpServletRequest request, NoteForm form, Model model) {
 
 		String referer;
 
-		this.sharesvc.toggleSharing(form.getId());
+		log.info("entering addSharedNote (Post-Method: /share)");
+
+		this.notesvc.toggleSharing(form.getId());
 
 		referer = request.getHeader("Referer");
- 		return "redirect:" + referer;
+		return "redirect:" + referer;
 	}
 
 }
-
