+++ src/main/java/de/thowl/tnt/web/TodoController.java
@@ -53,6 +53,7 @@ public class TodoController {
 	@RequestMapping(value = "/u/{username}/todo", method = RequestMethod.GET)
 	public String showTodoPage(@SessionAttribute(name = "token", required = false) AccessToken token,
 			@PathVariable("username") String username, Model model) {
+
 		log.info("entering showLoginPage (GET-Method: /u/{username}/todo)");
 
 		// Prevent unauthrised access / extend session
@@ -75,9 +76,9 @@ public class TodoController {
 			@SessionAttribute(name = "token", required = false) AccessToken token,
 			@PathVariable("username") String username, TaskForm form, Model model) {
 
-		log.info("entering doAddTask (POST-Method: /u/{}/todo)", username);
+		String referer;
 
-		String referer = request.getHeader("Referer");
+		log.info("entering doAddTask (POST-Method: /u/{}/todo)", username);
 
 		// Prevent unauthrised access / extend session
 		if (!this.authsvc.validateSession(token, username))
@@ -86,6 +87,7 @@ public class TodoController {
 		this.tasksvc.add(username, form.getTaskName(), form.getTaskContent(), form.getPriority(),
 				form.getDate(), form.getTime());
 
+		referer = request.getHeader("Referer");
 		return "redirect:" + referer;
 	}
 
@@ -99,9 +101,9 @@ public class TodoController {
 			@SessionAttribute(name = "token", required = false) AccessToken token,
 			@PathVariable("username") String username, TaskForm form, Model model) {
 
-		log.info("entering doAddTask (POST-Method: /u/{}/todo/done)", username);
+		String referer;
 
-		String referer = request.getHeader("Referer");
+		log.info("entering doAddTask (POST-Method: /u/{}/todo/done)", username);
 
 		// Prevent unauthrised access / extend session
 		if (!this.authsvc.validateSession(token, username))
@@ -109,6 +111,7 @@ public class TodoController {
 
 		this.tasksvc.toggleDone(form.getId());
 
+		referer = request.getHeader("Referer");
 		return "redirect:" + referer;
 	}
 
@@ -122,9 +125,9 @@ public class TodoController {
 			@SessionAttribute(name = "token", required = false) AccessToken token,
 			@PathVariable("username") String username, TaskForm form, Model model) {
 
-		log.info("entering doDeleteTask (DELETE-Method: /u/{}/todo)", username);
+		String referer;
 
-		String referer = request.getHeader("Referer");
+		log.info("entering doDeleteTask (DELETE-Method: /u/{}/todo)", username);
 
 		// Prevent unauthrised access / extend session
 		if (!this.authsvc.validateSession(token, username))
@@ -132,6 +135,7 @@ public class TodoController {
 
 		this.tasksvc.delete(form.getId());
 
+		referer = request.getHeader("Referer");
 		return "redirect:" + referer;
 	}
